Refactor and reorganize stakerConfig
