+++ packages/dappmanager/src/modules/stakerConfig/index.ts
@@ -1,2 +1,3 @@
 export { getStakerConfig } from "./get/getStakerConfig.js";
-export { setStakerConfig } from "./set/index.js";
+export { setStakerConfig } from "./set/setStakerConfig.js";
+export { getStakerConfigByNetwork } from "./getStakerConfigByNetwork.js";
